requirements.txt
country_converter==1.0.0
Pillow==10.0.0
pystray==0.19.4
requests==2.31.0
setuptools==68.0.0